requirements.txt
python-docx
pandas
pdf2docx
openpyxl